fix(64bit): canonicalize seed[3] to unsigned short + resolve 30 -Wsign-conversion sites (#11)
Phase 7 source fixes. signed<->unsigned conversions diverge on LP64;
-Wsign-conversion isolates exactly that class. Warn-only inventory reported
30 hits in 10 files, all in the olympia target (mapgen 0). 29 fixed
byte-identically; the seed fix deliberately corrects the saved value (golden
re-baselined in this commit).

seed[3] signedness mismatch (the headline defect): z.c:734 defines
`unsigned short seed[3]` (the erand48/drand48 state contract) but io.c:2459
and io.c:2583 declared `extern short seed[3]` (signed) -- a type-mismatch UB.
Both externs canonicalized to unsigned short, so seed[3] is unsigned
everywhere and the save routine (io.c:2602-2604, fprintf("%d", seed[i]))
writes the correct value.

This is the one deliberate golden change in the phase. The plan's
"golden-safe / identical bytes" premise was wrong: the fixture seeds span the
full 16-bit range (seed0=-3645) and the post-turn saved RNG state is routinely
>= 0x8000. The old signed extern wrote those negative (-28312); the correct
unsigned extern writes them positive (37224) -- same 16 bits, but the signed
text was wrong for an unsigned quantity. The golden `system` file is
re-baselined here to the correct unsigned seed values (the only file that
changed; one manifest line). The round-trip is unaffected -- load atoi()s then
truncates back into the 16-bit slot, recovering the exact bits -- so old saved
games still load correctly.

The other 29 hits were all int -> size_t/unsigned (every value provably >= 0),
fixed with explicit representation-preserving casts:
- qsort nmemb (the bulk): the element-typed list accessors (ilist_len,
  admits_len, item_ents_len, skill_ents_len) return int, feeding qsort's
  size_t nmemb. (size_t) cast at each call site in check.c/gm.c/input.c/
  perm.c/report.c/seed.c/swear.c/use.c, and once in the shared loop_known
  macro (loop.h:89), which cleared 6 sites at a stroke.
- my_malloc size (sout.c:31): (unsigned) cast on spaces_len + 1.
- array index (perm.c:451): (size_t) spaces_len - strlen(header).
- guard-allocator size store (z.c:260,:283): *((int *) p) = (int) size.

Debug + asan-ubsan golden gates both YES, asan/ubsan clean (exit 0).

+++ olympia/gm.c
@@ -55,7 +55,7 @@ gm_show_skill_use_counts(int pl)
 	}
 	next_skill;
 
-	qsort(l, ilist_len(l), sizeof(int), skill_use_comp);
+	qsort(l, (size_t) ilist_len(l), sizeof(int), skill_use_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -128,7 +128,7 @@ gm_show_skills_known(int pl)
 	}
 	next_skill;
 
-	qsort(l, ilist_len(l), sizeof(int), skills_known_comp);
+	qsort(l, (size_t) ilist_len(l), sizeof(int), skills_known_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -696,7 +696,7 @@ gm_land_stats(int pl)
 	}
 	next_loc;
 
-	qsort(l, ilist_len(l), sizeof(int), region_occupy_comp);
+	qsort(l, (size_t) ilist_len(l), sizeof(int), region_occupy_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -750,7 +750,7 @@ gm_faction_wealth(int pl)
 	}
 	next_player;
 
-	qsort(l, ilist_len(l), sizeof(int), wealth_list_comp);
+	qsort(l, (size_t) ilist_len(l), sizeof(int), wealth_list_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -803,7 +803,7 @@ gm_nobles_list(int pl)
 	}
 	next_player;
 
-	qsort(l, ilist_len(l), sizeof(int), nobles_list_comp);
+	qsort(l, (size_t) ilist_len(l), sizeof(int), nobles_list_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
